Day-1/Weather-Dashboard-Project/requirements.txt
boto3==1.26.137
python-dotenv==1.0.0
requests>=2.31.0
